pyproject.toml
[project]
name = "text-processing-tools-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "pyjwt>=2.10.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
